blogapp/serializers.py
from rest_framework import serializers
from .models import BlogPost, BlogCategory
class BlogPostSerializer(serializers.ModelSerializer):

    class Meta:
        model = BlogPost
        fields = '__all__'
        lookup_field = 'slug'

class BlogCategorySerializer(serializers.ModelSerializer):
    class Meta:
        model = BlogCategory
        fields = '__all__'